travelmind/backend/requirements.txt
google-genai
fastapi
uvicorn
python-dotenv
httpx
geopy
aiohttp